requirements.txt
Flask>=1.1.2
numpy>=1.19.5
tensorflow>=2.5.0
Pillow>=8.2.0
uuid>=1.30
shutilwhich>=1.1.0
gunicorn>=20.1.0